Move trajectory commands to new package
