+++ src/main/java/frc/robot/commands/trajectory/TrajectoryUtil.java
@@ -1,0 +1,4 @@
+package frc.robot.commands.trajectory;
+
+public class TrajectoryUtil {
+}
